Added methods to keep text mentioning shortcuts up to date, fixed potential issue with widget relationship tracking, hopefully got rid of signed/unsigned errors (Qt4/5 issue)

+++ apps/opencs/model/prefs/shortcutmanager.cpp
@@ -92,7 +92,7 @@ namespace CSMPrefs
         std::string output;
 
         // KeySequence
-        for (int i = 0; i < data.first.count(); ++i)
+        for (int i = 0; i < (int)data.first.count(); ++i)
         {
             if (data.first[i] & ModMask)
             {
